split the compressors in different files and extract the header logic to FilterHeaderUtils
