+++ colossus/src/main/scala/colossus/protocols/http/filters/FilterHeaderUtils.scala
@@ -1,0 +1,19 @@
+package colossus.protocols.http.filters
+
+import colossus.protocols.http.{ContentEncoding, HttpHeaders, HttpRequest, HttpRequestHead}
+import colossus.util.compress.{Compressor, GzipCompressor, ZCompressor}
+
+object FilterHeaderUtils {
+
+  def getCompressorFromHeader(head: HttpRequestHead, bufferedKb: Int): Option[(Compressor, ContentEncoding)] = {
+    head.headers.firstValue(HttpHeaders.AcceptEncoding).flatMap { header =>
+      if (header.contains(ContentEncoding.Gzip.value)) {
+        Some(new GzipCompressor(bufferedKb), ContentEncoding.Gzip)
+      } else if (header.contains(ContentEncoding.Deflate.value)) {
+        Some(new ZCompressor(bufferedKb), ContentEncoding.Deflate)
+      } else {
+        None
+      }
+    }
+  }
+}
