+++ colossus/src/main/scala/colossus/util/compress/Compressor.scala
@@ -1,0 +1,15 @@
+package colossus.util.compress
+
+import akka.util.ByteString
+
+trait Compressor {
+  def compress(bytes: ByteString): ByteString
+
+  def decompress(bytes: ByteString): ByteString
+}
+
+object NoCompressor extends Compressor {
+  def compress(bytes: ByteString): ByteString = bytes
+
+  def decompress(bytes: ByteString): ByteString = bytes
+}
